+++ sites/fast-component-explorer/app/fast-components/configs/fast-badge.ts
@@ -1,0 +1,40 @@
+import { fastComponentSchemas } from "@microsoft/site-utilities";
+import textSchema from "../../utilities/text.schema";
+import Guidance from "../../.tmp/badge/guidance";
+import { ComponentViewConfig } from "./data.props";
+
+export const fastBadgeId = "fast-badge";
+const fastBadgeConfig: ComponentViewConfig = {
+    schema: fastComponentSchemas[fastBadgeId],
+    guidance: Guidance,
+    scenarios: [
+        {
+            displayName: "Default",
+            dataDictionary: [
+                {
+                    root: {
+                        schemaId: fastBadgeId,
+                        data: {
+                            Slot: [
+                                {
+                                    id: "Slot",
+                                },
+                            ],
+                        },
+                    },
+                    Slot: {
+                        parent: {
+                            id: "root",
+                            dataLocation: "Slot",
+                        },
+                        schemaId: textSchema.id,
+                        data: "Badge",
+                    },
+                },
+                "root",
+            ],
+        },
+    ],
+};
+
+export default fastBadgeConfig;