+++ sites/fast-component-explorer/app/fast-components/configs/fast-slider.ts
@@ -1,0 +1,26 @@
+import { fastComponentSchemas } from "@microsoft/site-utilities";
+import Guidance from "../../.tmp/slider/guidance";
+import { ComponentViewConfig } from "./data.props";
+
+export const fastSliderId = "fast-slider";
+export const fastSliderLabelId = "fast-slider-label";
+const fastSliderConfig: ComponentViewConfig = {
+    schema: fastComponentSchemas[fastSliderId],
+    guidance: Guidance,
+    scenarios: [
+        {
+            displayName: "Default",
+            dataDictionary: [
+                {
+                    root: {
+                        schemaId: fastSliderId,
+                        data: {},
+                    },
+                },
+                "root",
+            ],
+        },
+    ],
+};
+
+export default fastSliderConfig;