+++ packages/react/fast-storybook-presets/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.4](https://github.com/Microsoft/fast-dna/compare/@microsoft/fast-storybook-presets@1.2.3...@microsoft/fast-storybook-presets@1.2.4) (2020-06-05)
+
+**Note:** Version bump only for package @microsoft/fast-storybook-presets
+
+
+
+
+
 ## [1.2.3](https://github.com/Microsoft/fast-dna/compare/@microsoft/fast-storybook-presets@1.2.2...@microsoft/fast-storybook-presets@1.2.3) (2020-05-18)
 
 **Note:** Version bump only for package @microsoft/fast-storybook-presets